Revert "Send request sebastian"
