+++ CODEOWNERS
@@ -1,2 +1,3 @@
 # default owners for non-specified
-*   @Jakeamorr
+# TODO I'm definitely going to change this after a chat with COE or the team leads or whoever
+*   @Ethan-Edmond @MarioAndresPortilla
