+++ CacheMoney/frontend/src/components/Account/CreateAccount.js
@@ -6,8 +6,6 @@ import axios from "axios";
 import { useState } from "react";
 import config from "../../config";
 import store from "../../store/Store.js";
-import { ToastContainer, toast } from 'react-toastify';
-import 'react-toastify/dist/ReactToastify.css';
 
 function CreateAccount(props) {
 	// local formData state
@@ -32,19 +30,6 @@ function CreateAccount(props) {
 					userId: store.getState().userReducer.userId,
 				},
 			})
-			.then(
-				//this isn't firing before the page re-renders. Either rework or remove.
-				toast.success('Account created!', {
-					position: "bottom-right",
-					autoClose: 2000,
-					hideProgressBar: true,
-					closeOnClick: true,
-					pauseOnHover: true,
-					draggable: true,
-					progress: undefined,
-				})
-			)
-
 			.catch((error) => console.error(`Error: ${error}`));
 	};
 
@@ -57,18 +42,15 @@ function CreateAccount(props) {
 	// what the submit button should do
 	const handleSubmit = (event) => {
 		postAccount(formData);
-		
-		props.handleClick(event)
-		
+		props.handleClick(event);
 		//navigate("/signin");
-
-			
+		alert(
+			"Account successfully created! Please return to the Accounts page to start using your new account."
+		);
 	};
 
 	return (
-
 		<div className="create-account-outer-container">
-		<ToastContainer/>
 			<div className="create-account-inner-container">
 				<div className="account_create_form">
 					<p className="account_create_form_header">Create Account</p>
