+++ CacheMoney/frontend/src/components/Transaction/Withdraw.js
@@ -4,8 +4,6 @@
 import axios from "axios";
 import config from "../../config";
 import store from "../../store/Store";
-import { ToastContainer, toast } from 'react-toastify';
-import 'react-toastify/dist/ReactToastify.css';
 
 function Withdraw(props) {
 	// post withdraw transaction
@@ -17,34 +15,7 @@ function Withdraw(props) {
 					userId: store.getState().userReducer.userId,
 				},
 			})
-
-			.then(
-				result=>{
-					result.status===200?
-					toast.success('Withdrawal successful', {
-						position: "bottom-right",
-						autoClose: 2000,
-						hideProgressBar: true,
-						closeOnClick: true,
-						pauseOnHover: true,
-						draggable: true,
-						progress: undefined,
-					}):toast.error('error')
-			}
-				)
-
-			.catch((error) => {
-				console.error(`Error: ${error}`)
-				toast.error('Withdrawal failed', {
-					position: "bottom-right",
-					autoClose: 2000,
-					hideProgressBar: true,
-					closeOnClick: true,
-					pauseOnHover: true,
-					draggable: true,
-					progress: undefined,
-					})
-			});
+			.catch((error) => console.error(`Error: ${error}`));
 	};
 
 	// what the submit button should do
@@ -69,7 +40,6 @@ function Withdraw(props) {
 
 	return (
 		<div className="withdraw-outer-container">
-			<ToastContainer />
 			<div className="withdraw-inner-container">
 				<div className="withdraw-form">
 					<p className="withdraw-form-header">Withdraw</p>
