--- buildspec.yml
@@ -1,42 +1,0 @@
-version: 0.2
-
-phases:
-  install:
-    runtime-versions:
-      java: corretto11
-  
-  pre_build:
-    commands:
-      # fact finding
-      - echo ======================================== PRE BUILD ========================================
-      - pwd
-      - ls -al
-
-      # move down to actual project directory
-      - cd CacheMoney/backend
-
-      # retrieve .properties from bucket
-      - aws s3 cp s3://revature-cachemoney/resources/application.properties src/main/resources/application.properties
-
-      # more fact finding
-      - ls -al
-      - ls -al src/main/resources/
-  
-  build:
-    commands:
-      # build Maven project to generate the WAR file
-      # skip tests to avoid Spring's wrath
-      - mvn package -DskipTests
-  
-  post_build:
-    commands:
-      # move WAR into the main directory
-      # name ROOT.war is associated with fix-path.config script
-      - mv target/backend-0.0.1-SNAPSHOT.war ../../ROOT.war
-      - cd ../../
-  
-artifacts:
-  # things to transmit to EBS for next step
-  files:
-    - ROOT.war
-    - .ebextensions/**/*
